Refactor code and add CondorMon magic command
